% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Birthday function}
\usage{
birthday(x)
}
\arguments{
\item{x}{The number of people in the group}
}
\value{
A probability scalar
}
\description{
Birthday function
}
\examples{
birthday(x = 20:25)
}
