% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data}
\format{
## `fire`
A data frame with 15 rows and 2 columns:
\describe{
  \item{distance}{Distance between the fire and the nearest fire station}
  \item{damage}{Amount of damage in thousands of dollars}
  ...
}
}
\source{
Data set from the book STATISTICS for Engineering and the Sciences SIXTH EDITION
}
\usage{
fire
}
\description{
A sample of 15 fires in a suburb. The amount of damage and the distance between the fire
and the nearest fire station are recorded for each fire
}
\keyword{datasets}
