% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{mybin}
\usage{
mybin(iter = 1000, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{sample size}

\item{p}{probability}
}
\value{
a table of relative frequencies and a barplot
}
\description{
mybin
}
\examples{
mybin(iter = 100, n = 10, p = 0.7)
}
