% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Plot a Normal Curve with a Shaded Probability Area}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{The mean}

\item{sigma}{The standard deviation}

\item{a}{The value to shade to}
}
\value{
A list of the mean, sd, cutoff, and probability
}
\description{
Plot a Normal Curve with a Shaded Probability Area
}
\examples{
myncurve(mu = 10, sigma = 2, a = 12)
}
